+++ config/dbConfig.js
@@ -1,12 +1,12 @@
 const sqlite3 = require("sqlite3").verbose();
 const buildSchemas = require("./schemas");
-
+const log = require("../logger/config/winston.config")
 // db = new sqlite3.Database(":memory:");
 const db = new sqlite3.Database(":memory:", (err) => {
   if (err) {
-    return console.error(err.message);
+    log.error(err.message);
   }
-  console.log("Connected to the in-memory SQlite database.");
+  log.info("Connected to the in-memory SQlite database.");
 });
 
 db.serialize(() => buildSchemas(db));
