+++ tests/app.test.js
@@ -1,11 +1,19 @@
-// const request = require("supertest");
-// const app = require("../index");
+const { expect } = require("chai");
+const request = require("supertest");
+const assert = require("assert");
+const app = require("../src/app");
 
-// describe("GET /health", () => {
-//   it("should return health", async (done) => {
-//     request(app)
-//       .get("/health")
-//       .expect("Content-Type", /text/)
-//       .expect(200, done);
-//   });
-// });
+describe("GET /health", () => {
+  it("should return health", async () => {
+    const res = await request(app)
+      .get("/health")
+      .expect("Content-Type", /text/);
+    assert.ok(res.statusCode === 200);
+  });
+  it("GET / should return 404", async () => {
+    const res = await request(app)
+      .get("/")
+      .expect(404);
+    assert.ok(res.statusCode === 404);
+  });
+});
