+++ tests/config/db.config.test.js
@@ -1,10 +1,11 @@
 const sqlite3 = require("sqlite3").verbose();
+const log = require("../../logger/config/winston.config");
 
 const db = new sqlite3.Database(":memory:", (err) => {
   if (err) {
-    console.error(err.message);
+    log.error(err.message);
   }
-  console.log("Connected to the in-memory SQlite test database.");
+  log.info("Connected to the in-memory SQlite test database.");
 });
 
 module.exports = db;
